+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## [0.1.4] 01.10.2024
 - rust_decimal dependency version upgrade to 1.36
-- new Compressibility quantity
+- new quantities: Compressibility and ReciprocalTemperature 
 ## [0.1.3] 23.05.2024
 - Updated CODATA 2022 Physical Constants
 ## [0.1.2] 16.05.2024